+++ kimchi/src/circuits/lookup/runtime_tables.rs
@@ -2,10 +2,7 @@
 //! The setup has to prepare for their presence using [`RuntimeTableCfg`].
 //! At proving time, the prover can use [`RuntimeTable`] to specify the actual tables.
 
-use crate::circuits::{
-    expr::{prologue::*, Column},
-    gate::CurrOrNext,
-};
+use crate::circuits::{berkeley_columns::Column, expr::prologue::*, gate::CurrOrNext};
 use ark_ff::Field;
 use serde::{Deserialize, Serialize};
