+++ kimchi/src/error.rs
@@ -69,10 +69,10 @@ pub enum VerifyError {
     IncorrectRuntimeProof,
 
     #[error("the evaluation for {0:?} is missing")]
-    MissingEvaluation(crate::circuits::expr::Column),
+    MissingEvaluation(crate::circuits::berkeley_columns::Column),
 
     #[error("the commitment for {0:?} is missing")]
-    MissingCommitment(crate::circuits::expr::Column),
+    MissingCommitment(crate::circuits::berkeley_columns::Column),
 }
 
 /// Errors that can arise when preparing the setup
